fix latest syntax for API nextjs 15+

+++ app/api/suara/route.ts
@@ -3,10 +3,10 @@ import Suara from "@/models/suara";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(request: NextRequest) {
-  const { nama, nomor, count } = await request.json();
-  await connect();
-  await Suara.create({ nama, nomor, count });
   try {
+    const { nama, nomor, count } = await request.json();
+    await connect();
+    await Suara.create({ nama, nomor, count });
     return NextResponse.json({ message: "berhasil dibuat" }, { status: 201 });
   } catch (error) {
     console.error(error);
@@ -15,16 +15,30 @@ export async function POST(request: NextRequest) {
 }
 
 export async function GET() {
-  await connect();
-  const suara = await Suara.find();
-  return NextResponse.json(suara);
+  try {
+    await connect();
+    const suara = await Suara.find();
+    return NextResponse.json(suara);
+  } catch (error) {
+    console.error(error);
+    return NextResponse.json(
+      { message: "Gagal mengambil data" },
+      { status: 500 },
+    );
+  }
 }
 
 export async function DELETE(request: NextRequest) {
-  const id = request.nextUrl.searchParams.get("id");
-  await connect();
-  await Suara.findByIdAndDelete(id);
   try {
+    const id = request.nextUrl.searchParams.get("id");
+    if (!id) {
+      return NextResponse.json(
+        { message: "Id tidak ditemukan" },
+        { status: 400 },
+      );
+    }
+    await connect();
+    await Suara.findByIdAndDelete(id);
     return NextResponse.json({ message: "Berhasil dihapus" }, { status: 200 });
   } catch (error) {
     console.error(error);
